fixing move TODO
