+++ source/common/http/BUILD
@@ -129,6 +129,7 @@ envoy_cc_library(
         "//include/envoy/http:conn_pool_interface",
         "//include/envoy/stats:timespan_interface",
         "//source/common/common:linked_object",
+        "//source/common/conn_pool:conn_pool_base_lib",
         "//source/common/stats:timespan_lib",
         "//source/common/upstream:upstream_lib",
     ],
